rename EXGL.h and EXGLContext.h to avoid conflicts
